rpm build - fixing some bugs when using private repos

+++ cicd/build/jenkins_build_rpm.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+set -x
+
+echo "(DEBUG) git version: " $(git --version)
 echo "(DEBUG) variables from upstream jenkin job (github-webhook):"
 # echo "(DEBUG)   \- REPOSITORY: $REPOSITORY" # (not used)
 # echo "(DEBUG)   \- EVENT: $EVENT" # (not used)
@@ -9,10 +12,13 @@ echo "(DEBUG)   \- RELEASE_TAG: $RELEASE_TAG"  # v3.211111, py3.220124
 echo "(DEBUG)   \- CRABSERVER_REPO: $CRABSERVER_REPO"  # dmwm, belforte, mapellidario, ...
 echo "(DEBUG)   \- WMCORE_REPO: $WMCORE_REPO"  # dmwm, belforte, mapellidario, ...
 echo "(DEBUG)   \- WMCORE_TAG: $WMCORE_TAG"  # <empty>, 1.5.7, ...
-# echo "(DEBUG)   \- BRANCH: $BRANCH" # (empty, not used)
 echo "(DEBUG)   \- PAYLOAD: $PAYLOAD"
 # example of PAYLOAD: https://dmapelli.web.cern.ch/public/crab/20220127/crabserver_github_release_payload_example.json
 echo "(DEBUG) end"
+
+if [[ -z $RELEASE_TAG ]]; then echo '$RELEASE_TAG' "is empty, exiting"; exit 1; fi
+if [[ -z $CRABSERVER_REPO ]]; then echo '$CRABSERVER_REPO' "is empty, exiting"; exit 1; fi
+if [[ -z $WMCORE_REPO ]]; then echo '$WMCORE_REPO' "is empty, exiting"; exit 1; fi
 
 #do a clean up
 docker system prune -af
@@ -20,27 +26,41 @@ docker system prune -af
 #clone directories
 git clone -b V00-33-XX https://github.com/cms-sw/pkgtools.git
 git clone https://github.com/cms-sw/cmsdist.git && cd cmsdist && git checkout comp_gcc630
-git clone https://github.com/dmwm/CRABServer.git
+git clone https://github.com/$CRABSERVER_REPO/CRABServer.git
+
+cd CRABServer
+git fetch --all --tags
+git checkout ${RELEASE_TAG}
+cd ..
 
 if [[ -n  ${PAYLOAD} ]]; then
-   #get CRABServer branch name from the payload, if the payload is not empty
+   #if the payload is not empty, get CRABServer branch name from the payload. 
    #the payload is not empty when this job is triggered by github-webhook,
    #if this job is launched manually, then the branch will have the value set
    #as input env variable.
    #paylod has a lot of information, but we are only interested in extracting this info: <...>"target_commitish": "python3"<...>
    #regex would extract 'python3' as a BRANCH name
    export BRANCH=$(echo "${PAYLOAD}" | grep -oP '(?<="target_commitish":\s")([^\s]+)(?=", ")')
+else
+   # when the payload is empty, it means that this job is launched manually.
+   # in this case, we extract the branch name from the $RELEASE_TAG
+   cd CRABServer
+   echo "(DEBUG) git branch -a --contains tags/$RELEASE_TAG: $(git branch -a --contains tags/$RELEASE_TAG)"
+   export BRANCH=$(git branch -a --contains tags/$RELEASE_TAG | sed "s/*//g" | sed "s/ //g" | grep origin | awk -F "/" '{print $3}')
+   cd ..
 fi
+if [[ -z $BRANCH ]]; then echo '$BRANCH' "is empty, exiting"; exit 1; fi
 echo "BRANCH=${BRANCH}" >> $WORKSPACE/properties_file
-
-cd CRABServer
-git checkout ${BRANCH}
-cd ..
+# ACHTUNG!
+# $BRANCH can contain only alfanumeric characters and the underscore "_". for example, the dash "-" is not allowed
+# otherwise, cms-build will fail and no rpm will be uploaded to the repository.
 
 #select the WMCore tag
 if [ ${WMCORE_REPO} == "dmwm" ]; then
    WMCORE_TAG=$(grep -oP "wmcver==\K.*" CRABServer/requirements.txt)
 fi
+if [[ -z $WMCORE_TAG ]]; then echo '$WMCORE_TAG' "is empty, exiting"; exit 1; fi
+
 
 echo "(DEBUG) env variables that could have been updated from the default:"
 echo "(DEBUG)   \- BRANCH (crabserver): $BRANCH"
